requirements.txt
ultralytics
fastapi
uvicorn[standard]
pillow
jinja2